Remove unnecessary cache interfaces
