+++ common/src/main/java/dk/kvalitetsit/itukt/common/repository/ClauseCache.java
@@ -2,8 +2,21 @@ package dk.kvalitetsit.itukt.common.repository;
 
 import dk.kvalitetsit.itukt.common.model.Clause;
 
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
-public interface ClauseCache {
-    Optional<Clause> getClause(String name);
+public class ClauseCache {
+    private final Map<String, Clause> nameToClauseMap;
+
+    public ClauseCache(List<Clause> clauses) {
+        this.nameToClauseMap = clauses.stream()
+                .collect(Collectors.toMap(Clause::name, Function.identity()));
+    }
+
+    public Optional<Clause> getClause(String name) {
+        return Optional.ofNullable(nameToClauseMap.get(name));
+    }
 }
