+++ common/src/test/java/dk/kvalitetsit/itukt/common/repository/ClauseCacheTest.java
@@ -1,4 +1,4 @@
-package dk.kvalitetsit.itukt.management.repository;
+package dk.kvalitetsit.itukt.common.repository;
 
 import dk.kvalitetsit.itukt.common.model.Clause;
 import org.junit.jupiter.api.Test;
@@ -7,12 +7,12 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ClauseCacheImplTest {
+class ClauseCacheTest {
 
     @Test
     void getClause_WhenClauseIsNotInCache_ReturnsEmptyOptional() {
         Clause existingClause = new Clause("test", null, null);
-        ClauseCacheImpl clauseCache = new ClauseCacheImpl(List.of(existingClause));
+        ClauseCache clauseCache = new ClauseCache(List.of(existingClause));
 
         var result = clauseCache.getClause("NonExistentClause");
 
@@ -23,7 +23,7 @@ class ClauseCacheImplTest {
     void getClause_WhenClauseIsInCache_ReturnsClause() {
         Clause existingClause1 = new Clause("test1", null, null);
         Clause existingClause2 = new Clause("test2", null, null);
-        ClauseCacheImpl clauseCache = new ClauseCacheImpl(List.of(existingClause1, existingClause2));
+        ClauseCache clauseCache = new ClauseCache(List.of(existingClause1, existingClause2));
 
         var result = clauseCache.getClause(existingClause1.name());
