+++ validation/src/main/java/dk/kvalitetsit/itukt/validation/repository/StamDataCache.java
@@ -1,7 +1,16 @@
 package dk.kvalitetsit.itukt.validation.repository;
 
+import java.util.Map;
 import java.util.Optional;
 
-public interface StamDataCache {
-    Optional<String> getClauseNameByDrugId(long drugId);
+public class StamDataCache {
+    private final Map<Long, String> drugIdToClauseNameMap;
+
+    public StamDataCache(Map<Long, String> drugIdToClauseNameMap) {
+        this.drugIdToClauseNameMap = drugIdToClauseNameMap;
+    }
+
+    public Optional<String> getClauseNameByDrugId(long drugId) {
+        return Optional.ofNullable(drugIdToClauseNameMap.get(drugId));
+    }
 }
